Enum wkhtmltopdf::Error

pub enum Error {
    IoError(Error),
    IllegalInit,
    NotInitialized,
    Blocked,
    ThreadMismatch(usize, usize),
    ConversionFailed(String),
    GlobalSettingFailure(String, String),
    ObjectSettingFailure(String, String),
}

The error type for wkhtmltopdf generation

Variants
IoError(Error)

Indicates an I/O error that occurred during PDF generation

IllegalInit

Indicates the wkhtmltopdf could be be initialized because it can only be initialized once per process (wkhtmltopdf limitation)

NotInitialized

Indicates that wkhtmltopdf has not yet been initialized in this process

Blocked

Indicates that wkhtmltopdf is blocked by another request within this process (wkhtmltopdf limitation)

ThreadMismatch(usize, usize)

Indicates that wkhtmltopdf was initialized on a different thread than this PDF generation atttempt (wkhtmltopdf limitation)

ConversionFailed(String)

Indicates that wkhtmltopdf conversion failed - internal error message comes directly from wkhtmltopdf

GlobalSettingFailure(String, String)

Indicates that wkhtmltopdf failed to set a particular global setting

ObjectSettingFailure(String, String)

Indicates that wkhtmltopdf failed to set a particular object setting