Module wkhtmltopdf::lowlevel

Low-level wkhtmltopdf without the raw pointers

This module abstracts away the raw pointers of wkhmtltox-sys while providing ownership and drop semantics necessary to safely use wkhtmltox-sys.

It is recommended to use the PdfBuilder build methods which manage all of these details, however, some usage scenarios (e.g. adding multiple objects to your PDF) may require using this lower-level module to achieve sufficient control.

Structs

PdfConverter
Safe wrapper for working with the wkhtmltopdf converter

PdfGlobalSettings
Safe wrapper for managing wkhtmltopdf global settings

PdfGuard
Handles initialization and deinitialization of wkhtmltopdf

PdfObjectSettings
Safe wrapper for managing wkhtmltopdf object settings

Functions

pdf_init
Initializes wkhtmltopdf