Struct wkhtmltopdf::lowlevel::PdfConverter

pub struct PdfConverter {
    // some fields omitted
}

Safe wrapper for working with the wkhtmltopdf converter

Methods

impl PdfConverter

fn add_page_object(&mut self, pdf_object: PdfObjectSettings, page: &str)

Adds a page object to the PDF by URL or local path to the page

This method will set/override the page object setting.

fn add_html_object(&mut self, pdf_object: PdfObjectSettings, html: &str)

Adds a page object to the PDF using provided HTML data

In general, this will result in ignoring the 'page' setting if added to this pdf_object. The exception is when html is an empty string, but app_page_object should be the preferred way to set the page setting.

fn convert<'a>(self) -> Result<PdfOutput<'a>>

Performs the HTML to PDF conversion

This method does not do any additional allocations of the output, so the PdfConverter will be owned by PdfOutput so that it is not dropped until the PdfOutput is dropped.